ThemeMenu to handle theme selection.  Added to Nav.  self generates list based on the themes declared in styles obj in constants
