+++ src/utils/constants.js
@@ -20,3 +20,5 @@ export const styles = {
     background: 'bg-space'
   }
 }
+
+export const themeList = Object.keys(styles)
